pyproject.toml
[project]
name = "pyenv"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "base58>=2.1.1",
    "boto3>=1.37.13",
    "flask>=3.1.0",
    "itsdangerous>=2.2.0",
    "pyjwt>=2.10.1",
    "requests>=2.32.3",
]
